[#280] kafka연결을 통한 재고의 대한 동시성 문제를 해결한다.

+++ src/main/java/project/trendpick_pro/domain/product/service/ProductService.java
@@ -5,9 +5,6 @@ import com.amazonaws.services.s3.AmazonS3;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.cache.annotation.CachePut;
-import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
@@ -26,13 +23,13 @@ import project.trendpick_pro.domain.common.base.rq.Rq;
 import project.trendpick_pro.domain.common.file.CommonFile;
 import project.trendpick_pro.domain.member.entity.Member;
 import project.trendpick_pro.domain.member.entity.RoleType;
-import project.trendpick_pro.domain.product.entity.Product;
-import project.trendpick_pro.domain.product.entity.dto.request.ProductSaveRequest;
-import project.trendpick_pro.domain.product.entity.dto.request.ProductSearchCond;
-import project.trendpick_pro.domain.product.entity.dto.response.ProductByRecommended;
-import project.trendpick_pro.domain.product.entity.dto.response.ProductListResponse;
-import project.trendpick_pro.domain.product.entity.dto.response.ProductListResponseBySeller;
-import project.trendpick_pro.domain.product.entity.dto.response.ProductResponse;
+import project.trendpick_pro.domain.product.entity.dto.ProductRequest;
+import project.trendpick_pro.domain.product.entity.product.Product;
+import project.trendpick_pro.domain.product.entity.product.dto.request.ProductSaveRequest;
+import project.trendpick_pro.domain.product.entity.product.dto.request.ProductSearchCond;
+import project.trendpick_pro.domain.product.entity.product.dto.response.*;
+import project.trendpick_pro.domain.product.entity.productOption.ProductOption;
+import project.trendpick_pro.domain.product.entity.productOption.dto.ProductOptionSaveRequest;
 import project.trendpick_pro.domain.product.exception.ProductNotFoundException;
 import project.trendpick_pro.domain.product.repository.ProductRepository;
 import project.trendpick_pro.domain.tags.favoritetag.entity.FavoriteTag;
@@ -41,6 +38,7 @@ import project.trendpick_pro.domain.tags.tag.entity.Tag;
 import project.trendpick_pro.domain.tags.tag.entity.type.TagType;
 import project.trendpick_pro.domain.tags.tag.service.TagService;
 import project.trendpick_pro.global.rsData.RsData;
+//import project.trendpick_pro.global.search.service.SearchService;
 
 import java.io.IOException;
 import java.util.*;
@@ -62,6 +60,7 @@ public class ProductService {
     private final TagService tagService;
 
     private final Rq rq;
+//    private final SearchService searchService;
 
     private final AmazonS3 amazonS3;
 
@@ -72,9 +71,12 @@ public class ProductService {
     private String filePath;
 
     @Transactional
-    public RsData<Long> register(ProductSaveRequest productSaveRequest, MultipartFile requestMainFile, List<MultipartFile> requestSubFiles) throws IOException {
+    public RsData<Long> register(ProductRequest request, MultipartFile requestMainFile, List<MultipartFile> requestSubFiles) throws IOException {
 
-        rq.CheckAdmin();
+        rq.checkAdmin();
+
+        ProductSaveRequest productSaveRequest = request.getRequest1();
+        ProductOptionSaveRequest optionSaveRequest = request.getRequest2();
 
         CommonFile mainFile = fileTranslator.translateFile(requestMainFile);
         List<CommonFile> subFiles = fileTranslator.translateFileList(requestSubFiles);
@@ -83,7 +85,7 @@ public class ProductService {
             mainFile.connectFile(subFile);
         }
 
-        Set<Tag> tags = new LinkedHashSet<>();  // 상품에 포함시킬 태크 선택하여 저장
+        Set<Tag> tags = new LinkedHashSet<>();
         for (String tagName : productSaveRequest.getTags()) {
             tags.add(new Tag(tagName));
         }
@@ -92,21 +94,27 @@ public class ProductService {
         SubCategory subCategory = subCategoryService.findByName(productSaveRequest.getSubCategory());
         Brand brand = brandService.findByName(productSaveRequest.getBrand());
 
-        Product product = Product.of(productSaveRequest, mainCategory, subCategory, brand, mainFile);
+        ProductOption productOption = ProductOption.of(optionSaveRequest);
+        Product product = Product.of(productSaveRequest, mainCategory, subCategory, brand, mainFile, productOption);
         product.addTag(tags);
 
-        productRepository.save(product);
-        return RsData.of("S-1", "상품 등록이 완료되었습니다.", product.getId());
+        Product saveProduct = productRepository.save(product);
+//        searchService.createProduct(saveProduct);
+        return RsData.of("S-1", "상품 등록이 완료되었습니다.", saveProduct.getId());
     }
 
     @Transactional
-    public RsData<Long> modify(Long productId, ProductSaveRequest productSaveRequest, MultipartFile requestMainFile, List<MultipartFile> requestSubFiles) throws IOException {
+    public RsData<Long> modify(Long productId, ProductRequest productRequest, MultipartFile requestMainFile, List<MultipartFile> requestSubFiles) throws IOException {
 
-        rq.CheckAdmin();
+        rq.checkAdmin();
 
-        Product product = productRepository.findById(productId).orElseThrow(() -> new ProductNotFoundException("존재하지 않는 상품입니다."));// 임시. 나중에 테스트
+        ProductSaveRequest productSaveRequest = productRequest.getRequest1();
+        ProductOptionSaveRequest optionSaveRequest = productRequest.getRequest2();
+
+        Product product = productRepository.findById(productId).orElseThrow(() -> new ProductNotFoundException("존재하지 않는 상품입니다."));
 
         product.getFile().deleteFile(amazonS3, bucket);
+        product.disconnectFile();
 
         CommonFile mainFile = fileTranslator.translateFile(requestMainFile);
         List<CommonFile> subFiles = fileTranslator.translateFileList(requestSubFiles);
@@ -115,51 +123,52 @@ public class ProductService {
             mainFile.connectFile(subFile);
         }
 
-        Set<Tag> tags = new LinkedHashSet<>();  // 상품에 포함시킬 태크 선택하여 저장
+        Set<Tag> tags = new LinkedHashSet<>();
         for (String tagName : productSaveRequest.getTags()) {
             tags.add(new Tag(tagName));
         }
 
         tagService.delete(product.getTags());
         product.modifyTag(tags);
-        product.update(productSaveRequest, mainFile);
+        product.update(productSaveRequest, optionSaveRequest, mainFile);
+//        searchService.modifyProduct(product);
         return RsData.of("S-1", "상품 수정 완료되었습니다.", product.getId());
     }
 
     @Transactional
     public void delete(Long productId) {
-        rq.CheckAdmin();
-        Product product = productRepository.findById(productId).orElseThrow(() -> new ProductNotFoundException("존재하지 않는 상품입니다."));// 임시. 나중에 테스트
+        rq.getAdmin();
+        Product product = productRepository.findById(productId).orElseThrow(() -> new ProductNotFoundException("존재하지 않는 상품입니다."));
         product.getFile().deleteFile(amazonS3, bucket);
+//        searchService.deleteProduct(product);
         productRepository.delete(product);
     }
 
     @Transactional
-    public ProductResponse show(Long productId) {
+    public ProductResponse getProduct(Long productId) {
 
-        Product product = productRepository.findById(productId).orElseThrow(() -> new ProductNotFoundException("존재하지 않는 상품입니다."));// 임시. 나중에 테스트
+        Product product = productRepository.findById(productId).orElseThrow(() -> new ProductNotFoundException("존재하지 않는 상품입니다."));
 
         if(rq.checkLogin()){
-            if(rq.CheckMemberHtml()){
+            if(rq.checkMember()){
                 updateFavoriteTag(product);
             }
         }
         return ProductResponse.of(product);
     }
 
-    public ProductListResponse getProduct(Long productId) {
+    public ProductListResponse getProducts(Long productId) {
         Product product = productRepository.findById(productId).orElseThrow(() -> new ProductNotFoundException("존재하지 않는 상품입니다."));
         return ProductListResponse.of(product);
     }
 
-
     private void updateFavoriteTag(Product product) {
-        Member member = rq.GetMember();
+        Member member = rq.getMember();
         if(member.getRole().equals(RoleType.MEMBER))
             favoriteTagService.updateTag(member, product, TagType.SHOW);
     }
 
-    public Page<ProductListResponse> showAll(int offset, String mainCategory, String subCategory) {
+    public Page<ProductListResponse> getProducts(int offset, String mainCategory, String subCategory) {
 
         ProductSearchCond cond = new ProductSearchCond(mainCategory, subCategory);
         PageRequest pageable = PageRequest.of(offset, 18);
@@ -167,7 +176,24 @@ public class ProductService {
         Page<ProductListResponse> listResponses = productRepository.findAllByCategoryId(cond, pageable);
 
         List<ProductListResponse> list = listResponses.getContent().stream()
-                .map(product -> getProduct(product.getId()))
+                .map(product -> getProducts(product.getId()))
+                .peek(product -> {
+                    String updatedMainFile = product.getMainFile();
+                    product.setMainFile(updatedMainFile);
+                }).toList();
+
+        return new PageImpl<>(list, pageable, listResponses.getTotalElements());
+    }
+
+    public Page<ProductListResponse> findAllByKeyword(String keyword, int offset) {
+
+        ProductSearchCond cond = new ProductSearchCond(keyword);
+        PageRequest pageable = PageRequest.of(offset, 18);
+
+        Page<ProductListResponse> listResponses = productRepository.findAllByKeyword(cond, pageable);
+
+        List<ProductListResponse> list = listResponses.getContent().stream()
+                .map(product -> getProducts(product.getId()))
                 .peek(product -> {
                     String updatedMainFile = product.getMainFile();
                     product.setMainFile(updatedMainFile);
@@ -244,17 +270,29 @@ public class ProductService {
         return RsData.of("S-1", "성공", productRepository.findAllBySeller(member.getBrand(), pageable));
     }
 
+    @Transactional
+    public void applyDiscount(Long productId, double discountRate) {
+        Product product = productRepository.findById(productId).orElseThrow(() -> new ProductNotFoundException("존재하지 않는 상품입니다."));
+        product.applyDiscount(discountRate);
+    }
+
     private ProductListResponse convertToProductListResponse(Product product) {
         return new ProductListResponse(
                 product.getId(),
                 product.getName(),
                 product.getBrand().getName(),
                 product.getFile().getFileName(),
-                product.getPrice()
+                product.getProductOption().getPrice(),
+                product.getDiscountRate(),
+                product.getDiscountedPrice()
         );
     }
 
     public Product findById(Long id) {
         return productRepository.findById(id).orElseThrow(() -> new ProductNotFoundException("존재하지 않는 상품입니다."));
+    }
+
+    public Product findByIdWithBrand(Long productId) {
+        return productRepository.findByIdWithBrand(productId);
     }
 }
